students_grades_in_a_class.py
Consider the following scenario about using if-elif-else statements:

Students in a class receive their grades as Pass/Fail. Scores of 60 or more (out of 100) mean that the grade is "Pass". For lower scores, the grade is "Fail". In addition, scores above 95 (not included) are graded as "Top Score".

Fill in the blanks in this function so that it returns the appropriate "Pass",  "Fail", or "Top Score" grade.