multiplication_table.py
Fill in the blanks to complete the function, which should output a multiplication table. The function accepts a variable “number” through its parameters. This “number” variable should be multiplied by the numbers 1 through 5, and printed in a format similar to “1x6=6” (“number” x “multiplier” = “result”). The code should also limit the “result” to not exceed 25. To satisfy these conditions, you will need to:

1. Initialize the “multiplier" variable with the starting value.
2. Complete the while loop condition.
3. Add an exit point for the loop.
4. Increment the “multiplier" variable inside the while loop.