splitting_the_words.py
Fill in the blank to complete the “string_words” function. This function should split up the words in the given “string” and return the number of words in the “string”.  Complete the string operation and method needed in this function so that a function call like "string_words("Hello, World")" will return the output "2".