sort_the_miles_driving_to_work.py
Consider the following scenario about using Python lists: 

Employees at a company shared the distance they drive to work (in miles) through an online survey. These distances were automatically added by Python to a list called “distances” in the order that each employee submitted their distance. Management wants the list to be sorted in the order of the longest distance to the shortest distance. 

Complete the function to sort the “distances” list. This function should:

1. sort the given “distances” list, passed through the function’s parameters; ; 
2. reverse the sort order so that it goes from the longest to the shortest distance;
3. return the modified “distances” list.